UBUNTU: [Debian] update-aufs.sh -- Revert apparently inadvertent changes
Some changes were made to this script and checked in as part of
adding in the aufs source. I suspect this was a mistake, and even
if not the script changes belong in a separate commit.

+++ debian/scripts/misc/update-aufs.sh
@@ -16,9 +16,14 @@ then
 fi
 
 clean=0
-rm -rf ${AUFS}
-git clone https://github.com/sfjro/aufs5-standalone.git ${AUFS}
-(cd ${AUFS}; git checkout -b aufs5.x-rcN remotes/origin/aufs5.10)
+if [ "$#" = 1 ]; then
+	AUFS="$1"
+else
+	clean=1
+	rm -rf ${AUFS}
+	git clone https://github.com/sfjro/aufs5-standalone.git ${AUFS}
+	(cd ${AUFS}; git checkout -b aufs5.x-rcN remotes/origin/aufs5.x-rcN)
+fi
 
 cp ${AUFS}/include/uapi/linux/aufs_type.h include/uapi/linux
 rsync -av ${AUFS}/fs/ fs/
